moved methods from BasePlayerClass to PlayerCharacter
