+++ app/src/main/java/com/philipgreen/dmwizard/PlayerCharacter.java
@@ -1,10 +1,18 @@
 package com.philipgreen.dmwizard;
 
+import android.util.Log;
+
+import com.philipgreen.dmwizard.data.BaseStats;
+import com.philipgreen.dmwizard.data.Skills;
+import com.philipgreen.dmwizard.dice.Dice;
 import com.philipgreen.dmwizard.playerClasses.BasePlayerClass;
 import com.philipgreen.dmwizard.races.BaseRaceClass;
 import com.philipgreen.dmwizard.weapons.abstractWeapons.BaseWeapon;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Created by pgreen on 8/11/16.
@@ -13,20 +21,268 @@ import java.util.ArrayList;
  * data such as classes, race, feats and equipment.
  */
 public class PlayerCharacter {
+    private static final String TAG = "PlayerCharacter";
     private ArrayList<BasePlayerClass> mClasses;
     private BaseRaceClass mRace;
     private ArrayList<BaseWeapon> mWeapons;
+
+    private int mStrength;
+    private int mDexterity;
+    private int mConstitution;
+    private int mIntelligence;
+    private int mWisdom;
+    private int mCharisma;
+
+    private BaseStats[] mSavingThrowProficiencies;
+    private int mStrengthSavingThrow;
+    private int mDexteritySavingThrow;
+    private int mConstitutionSavingThrow;
+    private int mIntelligenceSavingThrow;
+    private int mWisdomSavingThrow;
+    private int mCharismaSavingThrow;
+
+    private int mProficiencyBonus;
+    private int mInitiativeModifier;
+    private int mHitPoints;
+
+    private int mAcrobatics;
+    private int mAnimalHandling;
+    private int mArcana;
+    private int mAthletics;
+    private int mDeception;
+    private int mHistory;
+    private int mInsight;
+    private int mIntimidation;
+    private int mInvestigation;
+    private int mMedicine;
+    private int mNature;
+    private int mPerception;
+    private int mPerformance;
+    private int mPersuasion;
+    private int mReligion;
+    private int mSleightOfHand;
+    private int mStealth;
+    private int mSurvival;
+
+    // statically create and put values into ABILITY_MODIFIER_MAP
+    public static final Map<Integer, Integer> ABILITY_MODIFIER_MAP;
+
+    static {
+        Map<Integer, Integer> tempMap = new HashMap<Integer, Integer>();
+        tempMap.put(1, -5);
+        tempMap.put(2, -4);
+        tempMap.put(3, -4);
+        tempMap.put(4, -3);
+        tempMap.put(5, -3);
+        tempMap.put(6, -2);
+        tempMap.put(7, -2);
+        tempMap.put(8, -1);
+        tempMap.put(9, -1);
+        tempMap.put(10, 0);
+        tempMap.put(11, 0);
+        tempMap.put(12, 1);
+        tempMap.put(13, 1);
+        tempMap.put(14, 2);
+        tempMap.put(15, 2);
+        tempMap.put(16, 3);
+        tempMap.put(17, 3);
+        tempMap.put(18, 4);
+        tempMap.put(19, 4);
+        tempMap.put(20, 5);
+        ABILITY_MODIFIER_MAP = Collections.unmodifiableMap(tempMap);
+    }
 
     // Constructor for character with single class
     public PlayerCharacter(BasePlayerClass playerClass, BaseRaceClass playerRace) {
         this.mClasses.add(playerClass);
         this.mRace = playerRace;
+        this.mProficiencyBonus = playerClass.getProficiencyBonus();
+        this.mSavingThrowProficiencies = playerClass.getSavingThrowProficiencies();
+        this.mHitPoints = setHitPoints(playerClass.getClassLevel(), playerClass.getHitDie());
+
+        setSkills(playerClass.getSkillProficiencies());
+        setSavingThrows(mSavingThrowProficiencies);
     }
 
     // Constructor for character with multiple classes
     public PlayerCharacter(ArrayList<BasePlayerClass> playerClasses, BaseRaceClass playerRace) {
         this.mClasses = playerClasses;
         this.mRace = playerRace;
+
+        // TODO write class to handle multi-classing
+    }
+
+    private int setHitPoints(int classLevel, int hitDie) {
+        int totalHitPoints = 0;
+
+        for(int i = 0; i <= classLevel; i ++) {
+
+            // first level gets full hit points
+            if (i == 0) {
+                totalHitPoints += (hitDie + getConstitutionModifier());
+            } else {
+                totalHitPoints += (Dice.rollDie(hitDie) + getConstitutionModifier());
+            }
+        }
+
+        return totalHitPoints;
+    }
+
+    private void setSavingThrows(BaseStats[] savingThrowProficiencies) {
+        // Set proficient saving throws
+        for(BaseStats stat: savingThrowProficiencies) {
+            switch (stat) {
+                case STRENGTH:
+                    mStrengthSavingThrow = getStrengthModifier() + mProficiencyBonus;
+                    break;
+                case DEXTERITY:
+                    mDexteritySavingThrow = getDexterityModifier() + mProficiencyBonus;
+                    break;
+                case CONSTITUTION:
+                    mConstitutionSavingThrow = getConstitutionModifier() + mProficiencyBonus;
+                    break;
+                case INTELLIGENCE:
+                    mIntelligenceSavingThrow = getIntelligenceModifier() + mProficiencyBonus;
+                    break;
+                case WISDOM:
+                    mWisdomSavingThrow = getWisdomModifier() + mProficiencyBonus;
+                    break;
+                case CHARISMA:
+                    mCharismaSavingThrow = getCharismaModifier() + mProficiencyBonus;
+                    break;
+                default:
+                    Log.e(TAG, "Undefined enumeration constant: BaseStats: " + stat.toString());
+            }
+        }
+
+        // Whatever saving throws were not set should be set to character's stat modifier
+        if(mStrengthSavingThrow == 0) {
+            mStrengthSavingThrow = getStrengthModifier();
+        }
+        if(mDexteritySavingThrow == 0) {
+            mDexteritySavingThrow = getDexterityModifier();
+        }
+        if(mConstitutionSavingThrow == 0) {
+            mConstitutionSavingThrow = getConstitutionModifier();
+        }
+        if(mIntelligenceSavingThrow == 0) {
+            mIntelligenceSavingThrow = getIntelligenceModifier();
+        }
+        if(mWisdomSavingThrow == 0) {
+            mWisdomSavingThrow = getWisdomModifier();
+        }
+        if(mCharismaSavingThrow == 0) {
+            mCharismaSavingThrow = getCharismaModifier();
+        }
+
+    }
+
+    private void setSkills(Skills[] skillProficiencies) {
+        this.mAcrobatics = getDexterityModifier();
+        this.mAnimalHandling = getWisdomModifier();
+        this.mArcana = getIntelligenceModifier();
+        this.mAthletics = getStrengthModifier();
+        this.mDeception = getCharismaModifier();
+        this.mHistory = getIntelligenceModifier();
+        this.mInsight = getWisdomModifier();
+        this.mIntimidation = getCharismaModifier();
+        this.mInvestigation = getIntelligenceModifier();
+        this.mMedicine = getWisdomModifier();
+        this.mNature = getIntelligenceModifier();
+        this.mPerception = getWisdomModifier();
+        this.mPerformance = getCharismaModifier();
+        this.mPersuasion = getCharismaModifier();
+        this.mReligion = getIntelligenceModifier();
+        this.mSleightOfHand = getDexterityModifier();
+        this.mStealth = getDexterityModifier();
+        this.mSurvival = getWisdomModifier();
+        for (Skills skill : skillProficiencies) {
+            switch (skill) {
+                case ACROBATICS:
+                    this.mAcrobatics += mProficiencyBonus;
+                    break;
+                case ANIMAL_HANDLING:
+                    this.mAnimalHandling += mProficiencyBonus;
+                    break;
+                case ARCANA:
+                    this.mArcana += mProficiencyBonus;
+                    break;
+                case ATHLETICS:
+                    this.mAthletics += mProficiencyBonus;
+                    break;
+                case DECEPTION:
+                    this.mDeception += mProficiencyBonus;
+                    break;
+                case HISTORY:
+                    this.mHistory += mProficiencyBonus;
+                    break;
+                case INSIGHT:
+                    this.mInsight += mProficiencyBonus;
+                    break;
+                case INTIMIDATION:
+                    this.mIntimidation += mProficiencyBonus;
+                    break;
+                case INVESTIGATION:
+                    this.mInvestigation += mProficiencyBonus;
+                    break;
+                case MEDICINE:
+                    this.mMedicine += mProficiencyBonus;
+                    break;
+                case NATURE:
+                    this.mNature += mProficiencyBonus;
+                    break;
+                case PERCEPTION:
+                    this.mPerception += mProficiencyBonus;
+                    break;
+                case PERFORMANCE:
+                    this.mPerformance += mProficiencyBonus;
+                    break;
+                case PERSUASION:
+                    this.mPersuasion += mProficiencyBonus;
+                    break;
+                case RELIGION:
+                    this.mReligion += mProficiencyBonus;
+                    break;
+                case SLEIGHT_OF_HAND:
+                    this.mSleightOfHand += mProficiencyBonus;
+                    break;
+                case STEALTH:
+                    this.mStealth += mProficiencyBonus;
+                    break;
+                case SURVIVAL:
+                    this.mStealth += mProficiencyBonus;
+                    break;
+            }
+        }
+
+    }
+
+    ///////////////////////////////////////////////////
+    ////            ABILITY MODFIER GETTERS         ///
+    ///////////////////////////////////////////////////
+
+    public int getStrengthModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mStrength);
+    }
+
+    public int getDexterityModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mDexterity);
+    }
+
+    public int getConstitutionModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mConstitution);
+    }
+
+    public int getIntelligenceModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mIntelligence);
+    }
+
+    public int getWisdomModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mWisdom);
+    }
+
+    public int getCharismaModifier() {
+        return ABILITY_MODIFIER_MAP.get(this.mCharisma);
     }
 
 }
